numbers

+++ src/basic_types/numbers/numbers5.rs
@@ -1,6 +1,6 @@
 // Fix errors and panics to make it work
 fn main() {
-    let v1 = 251_u8 + 8;
-    let v2 = i8::checked_add(251, 8).unwrap();
+    let v1 = 251_u16 + 8;
+    let v2 = i16::checked_add(251, 8).unwrap();
     println!("{},{}", v1, v2);
 }
